requirements.txt
django==2.0.3
boto3==1.4.7
defusedxml==0.5.0
dj-database-url==0.4.2
django-ckeditor==5.4.0
django-storages==1.6.5
Pillow==5.0.0
python-decouple==3.1
requests==2.18.4
gunicorn==19.7.1
psycopg2==2.7.3.2